コメントを追加

+++ placefiles.sh
@@ -11,17 +11,52 @@ if [ ! -e $HOME/.emacs.d/ ]; then
     mkdir $HOME/.emacs.d/
 fi
 
-ln -s $currentdir/init.el init.el
-mv init.el $HOME/.emacs.d/
+echo "--- make init.el link ---"
+if [ ! -e $HOME/.emacs.d/init.el ]; then
+    ln -s $currentdir/init.el init.el
+    mv init.el $HOME/.emacs.d/
+    echo "OK."
+else
+    echo "File already exists."
+    echo "...skip"
+fi
+echo "-----------------------"
+echo ""
 
-ln -s $currentdir/Cask Cask
-mv Cask $HOME/.emacs.d/Cask
+echo "--- make Cask link ---"
+if [ ! -e $HOME/.emacs.d/Cask ]; then
+    ln -s $currentdir/Cask Cask
+    mv Cask $HOME/.emacs.d/Cask
+    echo "OK."
+else
+    echo "File already exists."
+    echo "...skip"
+fi
+echo "-----------------------"
+echo ""
 
-#ln -s $currentdir/.emacs.d .emacs.d
-#mv .emacs.d $HOME/
+echo "--- make .zshrc ---"
+if [ ! -e $HOME/.zshrc ]; then
+    ln -s $currentdir/.zshrc .zshrc
+    mv .zshrc $HOME/
+    echo "OK."
+else
+    echo "File already exists."
+    echo "...skip"
+fi
+echo "-----------------------"
+echo ""
 
-ln -s $currentdir/.zshrc .zshrc
-mv .zshrc $HOME/
+echo "--- make .tmux.conf link ---"
+if [ ! -e $HOME/.tmux.conf ]; then
+    ln -s $currentdir/.tmux.conf .tmux.conf
+    mv .tmux.conf $HOME/
+    echo "OK."
+else
+    echo "File already exists."
+    echo "...skip"
+fi
+echo "-----------------------"
+echo ""
 
-ln -s $currentdir/.tmux.conf .tmux.conf
-mv .tmux.conf $HOME/
+echo "\complete!/"
